move inputbindings and inputs resource to interface
