+++ crates/thetawave_interface/Cargo.toml
@@ -14,3 +14,4 @@ bevy_ecs_macros = { workspace = true }
 bevy_ecs = { workspace = true }
 bevy_reflect = { workspace = true }
 leafwing-input-manager = { workspace = true }
+bevy_input = { workspace = true }
