+++ crates/thetawave_interface/src/options/input.rs
@@ -1,6 +1,7 @@
-use bevy_ecs::component::Component;
+use bevy_ecs::{component::Component, system::Resource};
+use bevy_input::{gamepad::GamepadButtonType, keyboard::KeyCode};
 use bevy_reflect::Reflect;
-use leafwing_input_manager::Actionlike;
+use leafwing_input_manager::{prelude::InputMap, Actionlike};
 use serde::Deserialize;
 
 #[derive(Component)]
@@ -28,4 +29,31 @@ pub enum PlayerAction {
     BasicAttack,
     SpecialAttack,
     Pause,
+}
+
+#[derive(Deserialize)]
+pub struct InputBindings {
+    pub menu_keyboard: Vec<(KeyCode, MenuAction)>,
+    pub menu_gamepad: Vec<(GamepadButtonType, MenuAction)>,
+    pub player_keyboard: Vec<(KeyCode, PlayerAction)>,
+    pub player_gamepad: Vec<(GamepadButtonType, PlayerAction)>,
+}
+
+#[derive(Resource, Debug)]
+pub struct InputsResource {
+    pub menu: InputMap<MenuAction>,
+    pub player_keyboard: InputMap<PlayerAction>,
+    pub player_gamepad: InputMap<PlayerAction>,
+}
+
+impl InputsResource {
+    pub fn new(bindings: InputBindings) -> Self {
+        InputsResource {
+            menu: InputMap::new(bindings.menu_keyboard)
+                .insert_multiple(bindings.menu_gamepad)
+                .to_owned(),
+            player_keyboard: InputMap::new(bindings.player_keyboard),
+            player_gamepad: InputMap::new(bindings.player_gamepad),
+        }
+    }
 }
