+++ src/options/mod.rs
@@ -1,7 +1,10 @@
 //! `thetawave` player module
 use bevy::prelude::*;
 use leafwing_input_manager::prelude::InputManagerPlugin;
-use thetawave_interface::{options::input::MenuAction, states::AppStates};
+use thetawave_interface::{
+    options::input::{InputsResource, MenuAction},
+    states::AppStates,
+};
 
 mod display;
 mod input;
@@ -13,7 +16,7 @@ pub use self::display::{
 };
 use self::input::{
     get_input_bindings, read_menu_actions, spawn_menu_explorer_system,
-    spawn_player_controllers_system, InputsResource,
+    spawn_player_controllers_system,
 };
 
 pub struct OptionsPlugin;
